modify price type, change uint32 to uint64

+++ sqlchain/chain.go
@@ -1014,8 +1014,8 @@ func (c *Chain) SignBilling(low, high int32, unsigned *pt.BillingRequest) (
 	}
 
 	// Build map for later test
-	actualAmounts := make(map[proto.AccountAddress]uint32)
-	expectAmounts := make(map[proto.AccountAddress]uint32)
+	actualAmounts := make(map[proto.AccountAddress]uint64)
+	expectAmounts := make(map[proto.AccountAddress]uint64)
 
 	for _, v := range unsigned.Header.GasAmounts {
 		actualAmounts[v.AccountAddress] += v.GasAmount
